Massive refactor on macro scale
Massive refactor on macro scale, this includes:
- Splitting include/ and src/ into smaller folders.
- Putting files into their respective folders.
- Splitting big files into smaller ones for easier maintainance.
