+++ include/typecheck/Types.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Utilities.h"
+#include "util/Utilities.h"
 #include <cmath>
 #include <execution>
 #include <map>
