+++ include/util/LexicalContext.h
@@ -1,7 +1,7 @@
 #pragma once
 // A simple scoping class, doesn't differentiate between different names, like
 // variable name, func name and all that
-#include "Utilities.h"
+#include "util/Utilities.h"
 #include <functional>
 #include <memory>
 #include <ostream>
