+++ src/ast/CMakeLists.txt
@@ -1,0 +1,3 @@
+
+add_project_library(Ast)
+add_project_library(AstPrinterVisitor)
