+++ src/semantics/CMakeLists.txt
@@ -1,0 +1,2 @@
+
+add_project_library(ScopeGeneratorVisitor)
