+++ src/typecheck/CMakeLists.txt
@@ -1,0 +1,3 @@
+
+add_project_library(BiTypeChecker)
+add_project_library(Types)
